pyproject.toml
[tool.poetry]
name = "semantic-merkle-tree"
version = "0.1.0"
description = ""
authors = ["octavio <octavio.pavon@botman-ai.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-generativeai = "^0.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
